% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsing.R
\name{.delim_str_to_json}
\alias{.delim_str_to_json}
\title{Convert comma-separated string values into a JSON array string}
\usage{
.delim_str_to_json(str, delimiter = ",")
}
\arguments{
\item{str}{A string with values separated by commas}
}
\value{
The string containing a JSON array of the values
}
\description{
Convert comma-separated string values into a JSON array string
}
\examples{
.delim_str_to_json("foo, bar, baz")
}
