% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_portal_server.R
\name{get_portal_table}
\alias{get_portal_table}
\title{Get selected coloumns from a Synapse table.}
\usage{
get_portal_table(syn, table_id, cols)
}
\description{
Get selected coloumns from a Synapse table.
}
